+++ server/src/main/java/shinhan/mohaemoyong/server/adapter/deposit/DemandDepositApiAdapter.java
@@ -7,7 +7,7 @@ import org.springframework.web.client.RestTemplate;
 import shinhan.mohaemoyong.server.adapter.common.headerDto.RequestHeader;
 import shinhan.mohaemoyong.server.adapter.common.factory.HeaderFactory;
 import shinhan.mohaemoyong.server.adapter.deposit.dto.*;
-import shinhan.mohaemoyong.server.service.dto.InquireTransactionHistoryListRequestDto;
+import shinhan.mohaemoyong.server.service.financedto.InquireTransactionHistoryListRequestDto;
 
 
 @Slf4j
